Move RPC request type to medley
