+++ rolling-shutter/medley/rpc.go
@@ -1,0 +1,8 @@
+package medley
+
+type RPCRequest struct {
+	Version string      `json:"jsonrpc"`
+	Method  string      `json:"method,omitempty"`
+	Params  interface{} `json:"params,omitempty"`
+	ID      interface{} `json:"id,omitempty"`
+}
